Updated namespace. Moved test around
